Add cancellation context
This replaces internal channel/flag by passing a cancellation context.
At the same time it exploits that os/exec.Cmd in go >= 1.7 is context-
aware, i.e. os.Process.Kill() is performed when the context is canceled.

It adds context parameter, which makes it possible to integrate the
expect package into a possibly nested structure of cancellation contexts.

A second use case for this is to pass a context.WithTimeout/WithDeadline,
which will time out the command (useful for e.g. network-related calls).